+++ cli/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [5.12.0] - 2026-03-02
 
 ### Added
+- Added `sre job resume --job_id` for resubmitting standard jobs with checkpoint progress
 - Added support for insecure-skip-tls-verify in the `sre platform k8s create` command
 - Added bypassing of integration checks in developer mode
 - Added status hints to `sre rule update` command showing how to check status via `sre rulesource describe -rsid {rulesource_id} -cid "{customer_id}"`